Seperate declarations and implementations
