--- core/include/api/c_api_impl.hh
@@ -1,209 +1,0 @@
-/*
- * MIT License
- *
- * Copyright (c) 2017 Serge Zaitsev
- * Copyright (c) 2022 Steffen André Langnes
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-#ifndef WEBVIEW_API_C_API_IMPL_HH_LIB_HH
-#define WEBVIEW_API_C_API_IMPL_HH_LIB_HH
-
-#if defined(__cplusplus) && !defined(WEBVIEW_HEADER)
-#include "api/api_lib.hh"
-#include "errors/errors.h"
-#include "errors/errors.hh"
-#include "lib/macros.h"
-#include "lib/version.h"
-#include "strings/json.hh"
-#include "types/types.h"
-#include "webview_cc.hh"
-
-using namespace webview::_lib::_api;
-using namespace webview::types;
-
-WEBVIEW_API webview_t webview_create(int debug, void *wnd) {
-  using namespace webview::detail;
-  webview::webview *w{};
-  auto err = api_filter(
-      [=]() -> result<webview::webview *> {
-        return new webview::webview{static_cast<bool>(debug), wnd};
-      },
-      [&](webview::webview *w_) { w = w_; });
-  if (err == WEBVIEW_ERROR_OK) {
-    return w;
-  }
-  return nullptr;
-}
-
-WEBVIEW_API webview_error_t webview_destroy(webview_t w) {
-  using namespace webview::detail;
-  return api_filter([=]() -> noresult {
-    delete cast_to_webview(w);
-    return {};
-  });
-}
-
-WEBVIEW_API webview_error_t webview_run(webview_t w) {
-  using namespace webview::detail;
-  return api_filter([=] { return cast_to_webview(w)->run(); });
-}
-
-WEBVIEW_API webview_error_t webview_terminate(webview_t w) {
-  using namespace webview::detail;
-  return api_filter([=] { return cast_to_webview(w)->terminate(); });
-}
-
-WEBVIEW_API webview_error_t webview_dispatch(webview_t w,
-                                             void (*fn)(webview_t, void *),
-                                             void *arg) {
-  using namespace webview::detail;
-  if (!fn) {
-    return WEBVIEW_ERROR_INVALID_ARGUMENT;
-  }
-  return api_filter(
-      [=] { return cast_to_webview(w)->dispatch([=]() { fn(w, arg); }); });
-}
-
-WEBVIEW_API void *webview_get_window(webview_t w) {
-  using namespace webview::detail;
-  void *window = nullptr;
-  auto err = api_filter([=] { return cast_to_webview(w)->window(); },
-                        [&](void *value) { window = value; });
-  if (err == WEBVIEW_ERROR_OK) {
-    return window;
-  }
-  return nullptr;
-}
-
-WEBVIEW_API void *webview_get_native_handle(webview_t w,
-                                            webview_native_handle_kind_t kind) {
-  using namespace webview::detail;
-  void *handle{};
-  auto err = api_filter(
-      [=]() -> result<void *> {
-        auto *w_ = cast_to_webview(w);
-        switch (kind) {
-        case WEBVIEW_NATIVE_HANDLE_KIND_UI_WINDOW:
-          return w_->window();
-        case WEBVIEW_NATIVE_HANDLE_KIND_UI_WIDGET:
-          return w_->widget();
-        case WEBVIEW_NATIVE_HANDLE_KIND_BROWSER_CONTROLLER:
-          return w_->browser_controller();
-        default:
-          return error_info{WEBVIEW_ERROR_INVALID_ARGUMENT};
-        }
-      },
-      [&](void *handle_) { handle = handle_; });
-  if (err == WEBVIEW_ERROR_OK) {
-    return handle;
-  }
-  return nullptr;
-}
-
-WEBVIEW_API webview_error_t webview_set_title(webview_t w, const char *title) {
-  using namespace webview::detail;
-  if (!title) {
-    return WEBVIEW_ERROR_INVALID_ARGUMENT;
-  }
-  return api_filter([=] { return cast_to_webview(w)->set_title(title); });
-}
-
-WEBVIEW_API webview_error_t webview_set_size(webview_t w, int width, int height,
-                                             webview_hint_t hints) {
-  using namespace webview::detail;
-  return api_filter(
-      [=] { return cast_to_webview(w)->set_size(width, height, hints); });
-}
-
-WEBVIEW_API webview_error_t webview_navigate(webview_t w, const char *url) {
-  using namespace webview::detail;
-  if (!url) {
-    return WEBVIEW_ERROR_INVALID_ARGUMENT;
-  }
-  return api_filter([=] { return cast_to_webview(w)->navigate(url); });
-}
-
-WEBVIEW_API webview_error_t webview_set_html(webview_t w, const char *html) {
-  using namespace webview::detail;
-  if (!html) {
-    return WEBVIEW_ERROR_INVALID_ARGUMENT;
-  }
-  return api_filter([=] { return cast_to_webview(w)->set_html(html); });
-}
-
-WEBVIEW_API webview_error_t webview_init(webview_t w, const char *js) {
-  using namespace webview::detail;
-  if (!js) {
-    return WEBVIEW_ERROR_INVALID_ARGUMENT;
-  }
-  return api_filter([=] { return cast_to_webview(w)->init(js); });
-}
-
-WEBVIEW_API webview_error_t webview_eval(webview_t w, const char *js) {
-  using namespace webview::detail;
-  if (!js) {
-    return WEBVIEW_ERROR_INVALID_ARGUMENT;
-  }
-  return api_filter([=] { return cast_to_webview(w)->eval(js); });
-}
-
-WEBVIEW_API webview_error_t webview_bind(webview_t w, const char *name,
-                                         void (*fn)(const char *id,
-                                                    const char *req, void *arg),
-                                         void *arg) {
-  using namespace webview::detail;
-  if (!name || !fn) {
-    return WEBVIEW_ERROR_INVALID_ARGUMENT;
-  }
-  return api_filter([=] {
-    return cast_to_webview(w)->bind(
-        name,
-        [=](const std::string &seq, const std::string &req, void *arg_) {
-          fn(seq.c_str(), req.c_str(), arg_);
-        },
-        arg);
-  });
-}
-
-WEBVIEW_API webview_error_t webview_unbind(webview_t w, const char *name) {
-  using namespace webview::detail;
-  if (!name) {
-    return WEBVIEW_ERROR_INVALID_ARGUMENT;
-  }
-  return api_filter([=] { return cast_to_webview(w)->unbind(name); });
-}
-
-WEBVIEW_API webview_error_t webview_return(webview_t w, const char *id,
-                                           int status, const char *result) {
-  using namespace webview::detail;
-  if (!id || !result) {
-    return WEBVIEW_ERROR_INVALID_ARGUMENT;
-  }
-  return api_filter(
-      [=] { return cast_to_webview(w)->resolve(id, status, result); });
-}
-
-WEBVIEW_API const webview_version_info_t *webview_version(void) {
-  return &library_version_info;
-}
-
-#endif // defined(__cplusplus) && !defined(WEBVIEW_HEADER)
-#endif // WEBVIEW_API_C_API_IMPL_HH
