+++ core/include/detail/engine_base.hh
@@ -45,315 +45,135 @@ namespace detail {
 
 class engine_base {
 public:
-  engine_base(bool owns_window) : m_owns_window{owns_window} {}
-
+  /// Common internal API methods for all three Webview platform classes:
+  /// - cocoa_webkit
+  /// - gtk_webkitgtk
+  /// - win32_edge
+  engine_base(bool owns_window);
   virtual ~engine_base() = default;
-
-  noresult navigate(const std::string &url) {
-    if (url.empty()) {
-      return navigate_impl("about:blank");
-    }
-    return navigate_impl(url);
-  }
-
-  using sync_binding_t = std::function<std::string(std::string)>;
+  /// Internal API implementation of public \ref webview_navigate
+  noresult navigate(const std::string &url);
 
   // Synchronous bind
-  noresult bind(const std::string &name, sync_binding_t fn) {
-    auto wrapper = [this, fn](const std::string &id, const std::string &req,
-                              void * /*arg*/) { resolve(id, 0, fn(req)); };
-    return bind(name, wrapper, nullptr);
-  }
-
-  // Asynchronous bind
-  noresult bind(const std::string &name, binding_t fn, void *arg) {
-    // NOLINTNEXTLINE(readability-container-contains): contains() requires C++20
-    if (bindings.count(name) > 0) {
-      return error_info{WEBVIEW_ERROR_DUPLICATE};
-    }
-    bindings.emplace(name, binding_ctx_t(fn, arg));
-    replace_bind_script();
-    // Notify that a binding was created if the init script has already
-    // set things up.
-    eval("if (window.__webview__) {\n\
-window.__webview__.onBind(" +
-         json_escape(name) + ")\n\
-}");
-    return {};
-  }
-
-  noresult unbind(const std::string &name) {
-    auto found = bindings.find(name);
-    if (found == bindings.end()) {
-      return error_info{WEBVIEW_ERROR_NOT_FOUND};
-    }
-    bindings.erase(found);
-    replace_bind_script();
-    // Notify that a binding was created if the init script has already
-    // set things up.
-    eval("if (window.__webview__) {\n\
-window.__webview__.onUnbind(" +
-         json_escape(name) + ")\n\
-}");
-    return {};
-  }
-
+  noresult bind(const std::string &name, sync_binding_t fn);
+  /// Internal API implementation of public \ref webview_bind
+  noresult bind(const std::string &name, binding_t fn, void *arg);
+  /// Internal API implementation of public \ref webview_unbind
+  noresult unbind(const std::string &name);
+  /// Internal API implementation of public \ref webview_return
   noresult resolve(const std::string &id, int status,
-                   const std::string &result) {
-    // NOLINTNEXTLINE(modernize-avoid-bind): Lambda with move requires C++14
-    return dispatch(std::bind(
-        [id, status, this](std::string escaped_result) {
-          std::string js = "window.__webview__.onReply(" + json_escape(id) +
-                           ", " + std::to_string(status) + ", " +
-                           escaped_result + ")";
-          eval(js);
-        },
-        result.empty() ? "undefined" : json_escape(result)));
-  }
-
+                   const std::string &result);
+  /// Internal API implementation of public \ref webview_get_window
   result<void *> window() { return window_impl(); }
+  /// Internal API implementation part of public \ref webview_get_native_handle
   result<void *> widget() { return widget_impl(); }
+  /// Internal API implementation part of public \ref webview_get_native_handle
   result<void *> browser_controller() { return browser_controller_impl(); }
-  noresult run() { return run_impl(); }
-  noresult terminate() { return terminate_impl(); }
-  noresult dispatch(std::function<void()> f) { return dispatch_impl(f); }
-  noresult set_title(const std::string &title) { return set_title_impl(title); }
-
-  noresult set_size(int width, int height, webview_hint_t hints) {
-    auto res = set_size_impl(width, height, hints);
-    m_is_size_set = true;
-    return res;
-  }
-
-  noresult set_html(const std::string &html) { return set_html_impl(html); }
-
-  noresult init(const std::string &js) {
-    add_user_script(js);
-    return {};
-  }
-
-  noresult eval(const std::string &js) { return eval_impl(js); }
+  /// Internal API implementation of public \ref webview_run
+  noresult run();
+  /// Internal API implementation of public \ref webview_terminate
+  noresult terminate();
+  /// Internal API implementation of public \ref webview_dispatch
+  noresult dispatch(std::function<void()> f);
+  /// Internal API implementation of public \ref webview_set_title
+  noresult set_title(const std::string &title);
+  /// Internal API implementation of public \ref webview_set_size
+  noresult set_size(int width, int height, webview_hint_t hints);
+  /// Internal API implementation of public \ref webview_set_html
+  noresult set_html(const std::string &html);
+  /// Internal API implementation of public \ref webview_init
+  noresult init(const std::string &js);
+  /// Internal API implementation of public \ref webview_eval
+  noresult eval(const std::string &js);
 
 protected:
+  /// Platform specific implementation for \ref navigate
   virtual noresult navigate_impl(const std::string &url) = 0;
+  /// Platform specific implementation for \ref window
   virtual result<void *> window_impl() = 0;
+  /// Platform specific implementation for \ref widget
   virtual result<void *> widget_impl() = 0;
+  /// Platform specific implementation for \ref browser_controller
   virtual result<void *> browser_controller_impl() = 0;
+  /// Platform specific implementation for \ref run
   virtual noresult run_impl() = 0;
+  /// Platform specific implementation for \ref terminate
   virtual noresult terminate_impl() = 0;
+  /// Platform specific implementation for \ref dispatch
   virtual noresult dispatch_impl(std::function<void()> f) = 0;
+  /// Platform specific implementation for \ref set_title
   virtual noresult set_title_impl(const std::string &title) = 0;
+  /// Platform specific implementation for \ref set_size
   virtual noresult set_size_impl(int width, int height,
                                  webview_hint_t hints) = 0;
+  /// Platform specific implementation for \ref set_html
   virtual noresult set_html_impl(const std::string &html) = 0;
+  /// Platform specific implementation for \ref eval
   virtual noresult eval_impl(const std::string &js) = 0;
-
-  virtual user_script *add_user_script(const std::string &js) {
-    return std::addressof(*m_user_scripts.emplace(m_user_scripts.end(),
-                                                  add_user_script_impl(js)));
-  }
-
+  /// Adds a bound user function to Webview native code.
+  virtual user_script *add_user_script(const std::string &js);
+  /// Platform specific implementation to add a bound user JS function.
   virtual user_script add_user_script_impl(const std::string &js) = 0;
-
+  /// Platform specific implementation to remove all bound JS user functions from the Webview script.
   virtual void
   remove_all_user_scripts_impl(const std::list<user_script> &scripts) = 0;
-
+  /// Platform specific method to compare equality of two bound user functions.
   virtual bool are_user_scripts_equal_impl(const user_script &first,
                                            const user_script &second) = 0;
-
+  /// Replaces a bound user script in Webview native code.
   virtual user_script *replace_user_script(const user_script &old_script,
-                                           const std::string &new_script_code) {
-    remove_all_user_scripts_impl(m_user_scripts);
-    user_script *old_script_ptr{};
-    for (auto &script : m_user_scripts) {
-      auto is_old_script = are_user_scripts_equal_impl(script, old_script);
-      script = add_user_script_impl(is_old_script ? new_script_code
-                                                  : script.get_code());
-      if (is_old_script) {
-        old_script_ptr = std::addressof(script);
-      }
-    }
-    return old_script_ptr;
-  }
-
-  void replace_bind_script() {
-    if (m_bind_script) {
-      m_bind_script = replace_user_script(*m_bind_script, create_bind_script());
-    } else {
-      m_bind_script = add_user_script(create_bind_script());
-    }
-  }
-
-  void add_init_script(const std::string &post_fn) {
-    add_user_script(create_init_script(post_fn));
-    m_is_init_script_added = true;
-  }
-
-  std::string create_init_script(const std::string &post_fn) {
-    auto js = std::string{} + "(function() {\n\
-  'use strict';\n\
-  function generateId() {\n\
-    var crypto = window.crypto || window.msCrypto;\n\
-    var bytes = new Uint8Array(16);\n\
-    crypto.getRandomValues(bytes);\n\
-    return Array.prototype.slice.call(bytes).map(function(n) {\n\
-      var s = n.toString(16);\n\
-      return ((s.length % 2) == 1 ? '0' : '') + s;\n\
-    }).join('');\n\
-  }\n\
-  var Webview = (function() {\n\
-    var _promises = {};\n\
-    function Webview_() {}\n\
-    Webview_.prototype.post = function(message) {\n\
-      return (" +
-              post_fn + ")(message);\n\
-    };\n\
-    Webview_.prototype.call = function(method) {\n\
-      var _id = generateId();\n\
-      var _params = Array.prototype.slice.call(arguments, 1);\n\
-      var promise = new Promise(function(resolve, reject) {\n\
-        _promises[_id] = { resolve, reject };\n\
-      });\n\
-      this.post(JSON.stringify({\n\
-        id: _id,\n\
-        method: method,\n\
-        params: _params\n\
-      }));\n\
-      return promise;\n\
-    };\n\
-    Webview_.prototype.onReply = function(id, status, result) {\n\
-      var promise = _promises[id];\n\
-      if (result !== undefined) {\n\
-        try {\n\
-          result = JSON.parse(result);\n\
-        } catch (e) {\n\
-          promise.reject(new Error(\"Failed to parse binding result as JSON\"));\n\
-          return;\n\
-        }\n\
-      }\n\
-      if (status === 0) {\n\
-        promise.resolve(result);\n\
-      } else {\n\
-        promise.reject(result);\n\
-      }\n\
-    };\n\
-    Webview_.prototype.onBind = function(name) {\n\
-      if (window.hasOwnProperty(name)) {\n\
-        throw new Error('Property \"' + name + '\" already exists');\n\
-      }\n\
-      window[name] = (function() {\n\
-        var params = [name].concat(Array.prototype.slice.call(arguments));\n\
-        return Webview_.prototype.call.apply(this, params);\n\
-      }).bind(this);\n\
-    };\n\
-    Webview_.prototype.onUnbind = function(name) {\n\
-      if (!window.hasOwnProperty(name)) {\n\
-        throw new Error('Property \"' + name + '\" does not exist');\n\
-      }\n\
-      delete window[name];\n\
-    };\n\
-    return Webview_;\n\
-  })();\n\
-  window.__webview__ = new Webview();\n\
-})()";
-    return js;
-  }
-
-  std::string create_bind_script() {
-    std::string js_names = "[";
-    bool first = true;
-    for (const auto &binding : bindings) {
-      if (first) {
-        first = false;
-      } else {
-        js_names += ",";
-      }
-      js_names += json_escape(binding.first);
-    }
-    js_names += "]";
-
-    auto js = std::string{} + "(function() {\n\
-  'use strict';\n\
-  var methods = " +
-              js_names + ";\n\
-  methods.forEach(function(name) {\n\
-    window.__webview__.onBind(name);\n\
-  });\n\
-})()";
-    return js;
-  }
-
-  virtual void on_message(const std::string &msg) {
-    auto id = json_parse(msg, "id", 0);
-    auto name = json_parse(msg, "method", 0);
-    auto args = json_parse(msg, "params", 0);
-    auto found = bindings.find(name);
-    if (found == bindings.end()) {
-      return;
-    }
-    const auto &context = found->second;
-    dispatch([=] { context.call(id, args); });
-  }
-
-  virtual void on_window_created() { inc_window_count(); }
-
-  virtual void on_window_destroyed(bool skip_termination = false) {
-    if (dec_window_count() <= 0) {
-      if (!skip_termination) {
-        terminate();
-      }
-    }
-  }
-
-  // Runs the event loop until the currently queued events have been processed.
-  void deplete_run_loop_event_queue() {
-    bool done{};
-    dispatch([&] { done = true; });
-    run_event_loop_while([&] { return !done; });
-  }
-
-  // Runs the event loop while the passed-in function returns true.
+                                           const std::string &new_script_code);
+  /// Updates the JS `bind` script in the frontend window.
+  void replace_bind_script();
+  /// Adds the JS Webview script to the frontend window
+  void add_init_script(const std::string &post_fn);
+  /// Creates the JS Webview script for the frontend window
+  std::string create_init_script(const std::string &post_fn);
+  // Creates a `bind` JS script string for the frontend window.
+  std::string create_bind_script();
+  /// Handler for messages from the frontend window to the native Webview process.
+  virtual void on_message(const std::string &msg);
+  /// Handler to increment the browser window count
+  virtual void on_window_created();
+  /// Handler to decrement the browser window count
+  virtual void on_window_destroyed(bool skip_termination = false);
+  /// Various internal backend scenarios require platform event queue depletion.
+  /// This spins the platform event loop until queued events have been processed.
+  void deplete_run_loop_event_queue();
+  /// Various internal backend sccenarios require partial platform event queue depletion.
+  /// This spins the platform event loop until the callback function returns true.
   virtual void run_event_loop_while(std::function<bool()> fn) = 0;
-
-  void dispatch_size_default() {
-    if (!owns_window() || !m_is_init_script_added) {
-      return;
-    };
-    dispatch([this]() {
-      if (!m_is_size_set) {
-        set_size(m_initial_width, m_initial_height, WEBVIEW_HINT_NONE);
-      }
-    });
-  }
-
+  /// Adds the default window size event to the platform event queue.
+  /// This is guarded from execution if the user sets a preferred size.
+  void dispatch_size_default();
+  /// Explicitly toggle the guard for the setting of the default window size.
+  /// This may be needed to moderate behaviour during premature platform event queue depletion.
   void set_default_size_guard(bool guarded) { m_is_size_set = guarded; }
-
+  /// Gets a flag for whether the Webview window is embedded, or is owned by the user process.
   bool owns_window() const { return m_owns_window; }
 
 private:
-  static std::atomic_uint &window_ref_count() {
-    static std::atomic_uint ref_count{0};
-    return ref_count;
-  }
-
-  static unsigned int inc_window_count() { return ++window_ref_count(); }
-
-  static unsigned int dec_window_count() {
-    auto &count = window_ref_count();
-    if (count > 0) {
-      return --count;
-    }
-    return 0;
-  }
-
+  /// Keeps track of the number of platform window instances.
+  static std::atomic_uint &window_ref_count();
+  /// Increments the reference number of platform window instances.
+  static unsigned int inc_window_count();
+  /// Decrements the reference number of platform window instances.
+  static unsigned int dec_window_count();
+  /// A map of native promise callback functions.
   std::map<std::string, binding_ctx_t> bindings;
+  /// A reference to the currently active JS bound script in the platform window.
   user_script *m_bind_script{};
   std::list<user_script> m_user_scripts;
-
+  /// Flags if the JS Webview code has been sent to the platform window yet.
   bool m_is_init_script_added{};
+  /// Flags if the initial platform window size has been set.
+  /// This acts as a guard against the default size overriding a user set size
   bool m_is_size_set{};
+  /// Flags whether the Webview window is embedded or owned by the user process.
   bool m_owns_window{};
+  /// The default browser window width
   static const int m_initial_width = 640;
+  /// The default browser window height
   static const int m_initial_height = 480;
 };
