+++ core/include/detail/platform/windows/version.hh
@@ -33,10 +33,8 @@
 #ifdef _MSC_VER
 #pragma comment(lib, "version.lib")
 #endif
-#include "ntdll.hh"
 #include <array>
 #include <string>
-#include <vector>
 
 namespace webview {
 namespace detail {
@@ -64,93 +62,6 @@ struct version {
   static int compare_os_version(unsigned int major, unsigned int minor,
                                 unsigned int build);
 };
-
-template <typename T>
-std::array<unsigned int, 4>
-version::parse_version(const std::basic_string<T> &version) noexcept {
-  using iterator = typename std::basic_string<T>::const_iterator;
-  auto parse_component = [](iterator sb, iterator se) -> unsigned int {
-    try {
-      auto n = std::stol(std::basic_string<T>(sb, se));
-      return n < 0 ? 0 : n;
-    } catch (std::exception &) {
-      return 0;
-    }
-  };
-  auto end = version.end();
-  auto sb = version.begin(); // subrange begin
-  auto se = sb;              // subrange end
-  unsigned int ci = 0;       // component index
-  std::array<unsigned int, 4> components{};
-  while (sb != end && se != end && ci < components.size()) {
-    if (*se == static_cast<T>('.')) {
-      components[ci++] = parse_component(sb, se);
-      sb = ++se;
-      continue;
-    }
-    ++se;
-  }
-  if (sb < se && ci < components.size()) {
-    components[ci] = parse_component(sb, se);
-  }
-  return components;
-}
-
-template <typename T, std::size_t Length>
-std::array<unsigned int, 4>
-version::parse_version(const T (&version)[Length]) noexcept {
-  return parse_version(std::basic_string<T>(version, Length));
-}
-
-std::wstring
-version::get_file_version_string(const std::wstring &file_path) noexcept {
-  DWORD dummy_handle; // Unused
-  DWORD info_buffer_length =
-      GetFileVersionInfoSizeW(file_path.c_str(), &dummy_handle);
-  if (info_buffer_length == 0) {
-    return {};
-  }
-  std::vector<char> info_buffer;
-  info_buffer.reserve(info_buffer_length);
-  if (!GetFileVersionInfoW(file_path.c_str(), 0, info_buffer_length,
-                           info_buffer.data())) {
-    return {};
-  }
-  auto sub_block = L"\\StringFileInfo\\040904B0\\ProductVersion";
-  LPWSTR version = nullptr;
-  unsigned int version_length = 0;
-  if (!VerQueryValueW(info_buffer.data(), sub_block,
-                      reinterpret_cast<LPVOID *>(&version), &version_length)) {
-    return {};
-  }
-  if (!version || version_length == 0) {
-    return {};
-  }
-  return {version, version_length};
-}
-
-int version::compare_os_version(unsigned int major, unsigned int minor,
-                                unsigned int build) {
-  // Use RtlGetVersion both to bypass potential issues related to
-  // VerifyVersionInfo and manifests, and because both GetVersion and
-  // GetVersionEx are deprecated.
-  auto ntdll = native_library(L"ntdll.dll");
-  if (auto fn = ntdll.get(ntdll::RtlGetVersion())) {
-    RTL_OSVERSIONINFOW vi{};
-    vi.dwOSVersionInfoSize = sizeof(vi);
-    if (fn(&vi) != 0) {
-      return false;
-    }
-    if (vi.dwMajorVersion == major) {
-      if (vi.dwMinorVersion == minor) {
-        return static_cast<int>(vi.dwBuildNumber) - static_cast<int>(build);
-      }
-      return static_cast<int>(vi.dwMinorVersion) - static_cast<int>(minor);
-    }
-    return static_cast<int>(vi.dwMajorVersion) - static_cast<int>(major);
-  }
-  return false;
-}
 
 } // namespace windows
 } // namespace platform
