+++ core/src/api/webview_api_cc.hh
@@ -1,0 +1,8 @@
+#ifndef WEBVIEW_API_CC_HH
+#define WEBVIEW_API_CC_HH
+
+#if defined(__cplusplus) && !defined(WEBVIEW_HEADER)
+#include "api.cc"
+
+#endif // #if defined(__cplusplus) && !defined(WEBVIEW_HEADER)
+#endif // WEBVIEW_API_CC_HH
