+++ core/src/detail/platform/webview_detail_platform_cc.hh
@@ -1,0 +1,8 @@
+#ifndef WEBVIEW_DETAIL_PLATFORM_CC_HH
+#define WEBVIEW_DETAIL_PLATFORM_CC_HH
+
+#if defined(__cplusplus) && !defined(WEBVIEW_HEADER)
+#include "windows/webview_detail_platform_windows_cc.hh"
+
+#endif // defined(__cplusplus) && !defined(WEBVIEW_HEADER)
+#endif // WEBVIEW_DETAIL_PLATFORM_CC_HH
