+++ core/src/detail/platform/windows/webview_detail_platform_windows_cc.hh
@@ -1,0 +1,10 @@
+#ifndef WEBVIEW_DETAIL_PLATFORM_WINDWS_CC_HH
+#define WEBVIEW_DETAIL_PLATFORM_WINDWS_CC_HH
+
+#if defined(__cplusplus) && !defined(WEBVIEW_HEADER)
+#include "reg_key.cc"
+#include "version.cc"
+#include "webview2/loader.cc"
+
+#endif // defined(__cplusplus) && !defined(WEBVIEW_HEADER)
+#endif // WEBVIEW_DETAIL_PLATFORM_WINDWS_CC_HH
