+++ core/src/detail/threading/webview_detail_threading_cc.hh
@@ -1,0 +1,11 @@
+#ifndef WEBVIEW_DETAIL_THREADING_CC_HH
+#define WEBVIEW_DETAIL_THREADING_CC_HH
+
+#if defined(__cplusplus) && !defined(WEBVIEW_HEADER)
+//#include "atomic.cc"
+//#include "queue_thread_constructor.cc"
+//#include "resolve_thread_constructor.cc"
+//#include "threadsafe_lists.cc"
+
+#endif // defined(__cplusplus) && !defined(WEBVIEW_HEADER)
+#endif // WEBVIEW_DETAIL_THREADING_CC_HH
