+++ core/src/detail/webview_detail_cc.hh
@@ -1,0 +1,11 @@
+#ifndef WEBVIEW_DETAIL_CC_HH
+#define WEBVIEW_DETAIL_CC_HH
+
+#if defined(__cplusplus) && !defined(WEBVIEW_HEADER)
+#include "engine_base.cc"
+//#include "engine_queue.cc"
+#include "platform/webview_detail_platform_cc.hh"
+#include "threading/webview_detail_threading_cc.hh"
+
+#endif // defined(__cplusplus) && !defined(WEBVIEW_HEADER)
+#endif // WEBVIEW_DETAIL_CC_HH
