+++ core/src/log/webview_log_cc.hh
@@ -1,0 +1,8 @@
+#ifndef WEBVIEW_LOG_CC_HH
+#define WEBVIEW_LOG_CC_HH
+
+#if defined(__cplusplus) && !defined(WEBVIEW_HEADER)
+//#include "trace_log.cc"
+
+#endif // defined(__cplusplus) && !defined(WEBVIEW_HEADER)
+#endif // WEBVIEW_LOG_CC_HH
