+++ core/src/strings/webview_strings_cc.hh
@@ -1,0 +1,8 @@
+#ifndef WEBVIEW_STRINGS_CC_HH
+#define WEBVIEW_STRINGS_CC_HH
+
+#if defined(__cplusplus) && !defined(WEBVIEW_HEADER)
+//#include "string_maker.cc"
+
+#endif // defined(__cplusplus) && !defined(WEBVIEW_HEADER)
+#endif // WEBVIEW_STRINGS_CC_HH
