+++ core/src/tests/webview_tests_cc.hh
@@ -1,0 +1,9 @@
+#ifndef WEBVIEW_TESTS_CC_HH
+#define WEBVIEW_TESTS_CC_HH
+
+#if defined(__cplusplus) && !defined(WEBVIEW_HEADER)
+//#include "test_helper.cc"
+//#include "test_templates.cc"
+
+#endif // defined(__cplusplus) && !defined(WEBVIEW_HEADER)
+#endif // WEBVIEW_TESTS_CC_HH
