+++ core/src/webview_cc.hh
@@ -1,0 +1,12 @@
+#ifndef WEBVIEW_CC_HH
+#define WEBVIEW_CC_HH
+
+#if defined(__cplusplus) && !defined(WEBVIEW_HEADER)
+#include "api/webview_api_cc.hh"
+#include "detail/webview_detail_cc.hh"
+#include "log/webview_log_cc.hh"
+#include "strings/webview_strings_cc.hh"
+#include "tests/webview_tests_cc.hh"
+
+#endif // defined(__cplusplus) && !defined(WEBVIEW_HEADER)
+#endif // WEBVIEW_CC_HH
